refactor pr
